Docker

+++ BirthdayBot.BLL/Commands/Notes/AddNote.cs
@@ -12,6 +12,8 @@ using RapidBots.Types.Core;
 using System.Threading.Tasks;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.ReplyMarkups;
+using BirthdayBot.Core.Const;
+using Microsoft.EntityFrameworkCore;
 
 namespace BirthdayBot.BLL.Commands.Notes
 {
@@ -33,28 +35,39 @@ namespace BirthdayBot.BLL.Commands.Notes
             var actionsManager = actionScope.ServiceProvider.GetService<ActionManager>();
             var resources = actionScope.ServiceProvider.GetService<IStringLocalizer<SharedResources>>();
 
-            TUser dbUser = user as TUser ?? await repository.GetAsync<TUser>(false, u => u.Id == update.CallbackQuery.From.Id);
-            dbUser.CurrentStatus = actionsManager.FindInputStatusByType<NoteTitleInput>();
-            var newNote = new Note();
-            dbUser.MiddlewareData = JsonConvert.SerializeObject(newNote);
-            await repository.UpdateAsync(dbUser);
-
-            try{await botClient.AnswerCallbackQueryAsync(update.CallbackQuery.Id);}catch{}
-            try
+            TUser dbUser = (user as TUser) ?? await repository.GetAsync<TUser>(false, u => u.Id == update.CallbackQuery.From.Id, include: u => u.Include(x => x.Notes));
+            if (dbUser?.Notes == null)
             {
-                await botClient.DeleteMessageAsync(update.CallbackQuery.Message.Chat.Id, update.CallbackQuery.Message.MessageId);
+                await repository.LoadCollectionAsync(dbUser, x => x.Notes);
             }
-            catch
-            { }
 
-            // Output
+            if (dbUser.Notes.Count < Limitations.NotesLimit)
+            {
+                dbUser.CurrentStatus = actionsManager.FindInputStatusByType<NoteTitleInput>();
+                var newNote = new Note();
+                dbUser.MiddlewareData = JsonConvert.SerializeObject(newNote);
+                await repository.UpdateAsync(dbUser);
 
-            KeyboardButton backBut = new KeyboardButton() { Text = resources["BACK_BUTTON"] };
+                try { await botClient.AnswerCallbackQueryAsync(update.CallbackQuery.Id); } catch { }
+                try
+                {
+                    await botClient.DeleteMessageAsync(update.CallbackQuery.Message.Chat.Id, update.CallbackQuery.Message.MessageId);
+                }
+                catch
+                { }
 
-            dbUser.CurrentStatus = actionsManager.FindInputStatusByType<NoteTitleInput>();
-            await repository.UpdateAsync(dbUser);
-            await botClient.SendTextMessageAsync(update.Message?.Chat?.Id ?? update.CallbackQuery.Message.Chat.Id, resources["NOTE_TITLE_INPUT"], replyMarkup: new ReplyKeyboardMarkup(backBut) { ResizeKeyboard = true }, parseMode: Telegram.Bot.Types.Enums.ParseMode.Html);
+                // Output
 
+                KeyboardButton backBut = new KeyboardButton() { Text = resources["BACK_BUTTON"] };
+
+                dbUser.CurrentStatus = actionsManager.FindInputStatusByType<NoteTitleInput>();
+                await repository.UpdateAsync(dbUser);
+                await botClient.SendTextMessageAsync(update.Message?.Chat?.Id ?? update.CallbackQuery.Message.Chat.Id, resources["NOTE_TITLE_INPUT"], replyMarkup: new ReplyKeyboardMarkup(backBut) { ResizeKeyboard = true }, parseMode: Telegram.Bot.Types.Enums.ParseMode.Html);
+            }
+            else
+            {
+                await botClient.AnswerCallbackQueryAsync(update.CallbackQuery.Id, resources["NOTES_LIMIT"], true);
+            }
         }
     }
 }
