+++ BirthdayBot.BLL/Commands/Notes/AddNoteFromPersonal.cs
@@ -13,6 +13,8 @@ using System.Threading.Tasks;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.ReplyMarkups;
 using RapidBots.Extensions;
+using BirthdayBot.Core.Const;
+using Microsoft.EntityFrameworkCore;
 
 namespace BirthdayBot.BLL.Commands.Notes
 {
@@ -37,32 +39,43 @@ namespace BirthdayBot.BLL.Commands.Notes
             var actionsManager = actionScope.ServiceProvider.GetService<ActionManager>();
             var resources = actionScope.ServiceProvider.GetService<IStringLocalizer<SharedResources>>();
 
-            TUser dbUser = user as TUser ?? await repository.GetAsync<TUser>(false, u => u.Id == update.CallbackQuery.From.Id);
-            dbUser.CurrentStatus = actionsManager.FindInputStatusByType<NoteTitleInput>();
-            var newNote = new Note() { Title = title};
-            newNote.Title = title;
+            TUser dbUser = (user as TUser) ?? await repository.GetAsync<TUser>(false, u => u.Id == update.CallbackQuery.From.Id, include: u => u.Include(x => x.Notes));
 
-            await botClient.SendTextMessageAsync(update.Message?.Chat?.Id ?? update.CallbackQuery.Message.Chat.Id, resources["TRANSFERED_NOTE_CREATION", title], parseMode: Telegram.Bot.Types.Enums.ParseMode.Html);
-
-            dbUser.MiddlewareData = JsonConvert.SerializeObject(newNote);
-            await repository.UpdateAsync(dbUser);
-
-            try { await botClient.AnswerCallbackQueryAsync(update.CallbackQuery.Id); } catch { }
-            try
+            if (dbUser?.Notes == null)
             {
-                await botClient.DeleteMessageAsync(update.CallbackQuery.Message.Chat.Id, update.CallbackQuery.Message.MessageId);
+                await repository.LoadCollectionAsync(dbUser, x => x.Notes);
             }
-            catch
-            { }
+            if(dbUser.Notes.Count < Limitations.NotesLimit)
+            {
+                dbUser.CurrentStatus = actionsManager.FindInputStatusByType<NoteTitleInput>();
+                var newNote = new Note() { Title = title };
+                newNote.Title = title;
 
-            // Output
+                await botClient.SendTextMessageAsync(update.Message?.Chat?.Id ?? update.CallbackQuery.Message.Chat.Id, resources["TRANSFERED_NOTE_CREATION", title], parseMode: Telegram.Bot.Types.Enums.ParseMode.Html);
 
-            KeyboardButton backBut = new KeyboardButton() { Text = resources["BACK_BUTTON"] };
+                dbUser.MiddlewareData = JsonConvert.SerializeObject(newNote);
+                await repository.UpdateAsync(dbUser);
 
-            dbUser.CurrentStatus = actionsManager.FindInputStatusByType<NoteYearInput>();
-            await repository.UpdateAsync(dbUser);
-            await botClient.SendTextMessageAsync(update.Message?.Chat?.Id ?? update.CallbackQuery.Message.Chat.Id, resources["NOTE_YEAR_INPUT"], parseMode: Telegram.Bot.Types.Enums.ParseMode.Html);
+                try { await botClient.AnswerCallbackQueryAsync(update.CallbackQuery.Id); } catch { }
+                try
+                {
+                    await botClient.DeleteMessageAsync(update.CallbackQuery.Message.Chat.Id, update.CallbackQuery.Message.MessageId);
+                }
+                catch
+                { }
 
+                // Output
+
+                KeyboardButton backBut = new KeyboardButton() { Text = resources["BACK_BUTTON"] };
+
+                dbUser.CurrentStatus = actionsManager.FindInputStatusByType<NoteYearInput>();
+                await repository.UpdateAsync(dbUser);
+                await botClient.SendTextMessageAsync(update.Message?.Chat?.Id ?? update.CallbackQuery.Message.Chat.Id, resources["NOTE_YEAR_INPUT"], parseMode: Telegram.Bot.Types.Enums.ParseMode.Html);
+            }
+            else
+            {
+                await botClient.AnswerCallbackQueryAsync(update.CallbackQuery.Id, resources["NOTES_LIMIT"], true);
+            }
         }
     }
 }
